+++ lib/accounts_list/view/accounts_list_page.dart
@@ -52,7 +52,7 @@ class AccountsListView extends StatelessWidget {
             final themeState = context.watch<ThemeCubit>().state;
             return Scaffold(
               appBar: AppBar(
-                title: Text('Accounts', style: TextStyle(fontSize: 36.sp)),
+                title: Text('Accounts', style: TextStyle(fontSize: 30.sp)),
                 leading: IconButton(
                   icon: Icon(Icons.close),
                   onPressed: () {
@@ -62,51 +62,48 @@ class AccountsListView extends StatelessWidget {
               ),
               body: Column(
                 children: [
-                  SizedBox(
-                    height: 10,
-                  ),
+                  SizedBox(height: 10),
                   Expanded(
                     child: ListView.builder(
                       itemCount: state.accountList.length,
                       itemBuilder: (context, index) {
                         final account = state.accountList[index];
                         return Card(
-                          elevation: Theme.of(context).cardTheme.elevation,
-                          child: ListTile(
-                            title: Row(
-                              children: [
-                                Text(
-                                  account.extendName(),
-                                  style: TextStyle(
-                                      overflow: TextOverflow.ellipsis,
-                                      color: BudgetTheme.isDarkMode(context)
-                                          ? Colors.white
-                                          : Colors.black,
-                                      fontSize: 28.sp,
-                                      fontWeight: FontWeight.w500),
-                                ),
-                                Expanded(child: Container()),
-                                FaIcon(
+                          child: Row(
+                            children: [
+                              IconButton(
+                                icon: Icon(Icons.highlight_remove,
+                                    color:
+                                        Theme.of(context).colorScheme.error),
+                                onPressed: () {
+                                  context
+                                      .read<AccountsCubit>()
+                                      .onAccountDeleted(account);
+                                },
+                              ),
+                              Text(
+                                overflow: TextOverflow.ellipsis,
+                                account.extendName(),
+                                style: TextStyle(
                                     color: BudgetTheme.isDarkMode(context)
                                         ? Colors.white
                                         : Colors.black,
-                                    IconData(account.category.iconCode,
-                                        fontFamily: 'FontAwesomeSolid')),
-                              ],
-                            ),
-                            leading: IconButton(
-                              icon: Icon(Icons.highlight_remove,
-                                  color: Theme.of(context).colorScheme.error),
-                              onPressed: () {
-                                context
-                                    .read<AccountsCubit>()
-                                    .onAccountDeleted(account);
-                              },
-                            ),
-                            trailing: Icon(Icons.chevron_right),
-                            onTap: () {
-                              context.push('/accounts-list/edit/${account.id}');
-                            },
+                                    fontSize: 22.sp,
+                                    fontWeight: FontWeight.w500),
+                              ),
+                              Spacer(),
+                              FaIcon(
+                                  color: BudgetTheme.isDarkMode(context)
+                                      ? Colors.white
+                                      : Colors.black,
+                                  IconData(account.category.iconCode,
+                                      fontFamily: 'FontAwesomeSolid')),
+                              IconButton(
+                                onPressed: () => context.push(
+                                    '/accounts-list/edit/${account.id}'),
+                                icon: Icon(Icons.chevron_right),
+                              )
+                            ],
                           ),
                         );
                       },