+++ lib/backup/view/view.dart
@@ -1,0 +1,2 @@
+export 'backup_page.dart';
+export 'widgets/widgets.dart';