+++ lib/navigation/router.dart
@@ -1,5 +1,6 @@
 import 'package:qruto_budget/accounts_list/repository/account_repository.dart';
 import 'package:qruto_budget/accounts_list/view/accounts_list_page.dart';
+import 'package:qruto_budget/backup/view/backup_page.dart';
 import 'package:qruto_budget/categories/category_edit/view/category_edit_dialog.dart';
 import 'package:qruto_budget/categories/repository/category_repository.dart';
 import 'package:qruto_budget/categories/view/categories_page.dart';
@@ -24,8 +25,6 @@ GoRouter get router => _router;
 
 final GlobalKey<NavigatorState> _rootNavigatorKey =
     GlobalKey<NavigatorState>(debugLabel: 'root');
-final GlobalKey<NavigatorState> _buildingsNavigatorKey =
-    GlobalKey<NavigatorState>(debugLabel: 'buildingsNav');
 
 /// The route configuration.
 final GoRouter _router = GoRouter(
@@ -190,6 +189,12 @@ final List<RouteBase> _individualRoutes = [
           },
         ),
       ]),
+  GoRoute(
+    path: '/backup',
+    builder: (BuildContext context, GoRouterState state) {
+      return BackupPage();
+    },
+  ),
   GoRoute(
     path: '/settings',
     builder: (BuildContext context, GoRouterState state) {