+++ lib/transfer/view/transfer_form.dart
@@ -55,7 +55,6 @@ class TransferForm extends StatelessWidget {
 class _SubmitButton extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final colors = context.read<ThemeCubit>().state;
     return BlocBuilder<TransferBloc, TransferState>(
       builder: (context, state) {
         final themeState = context.watch<ThemeCubit>().state;
@@ -80,7 +79,7 @@ class _SubmitButton extends StatelessWidget {
                         child: Text(
                           'Submit',
                           style: TextStyle(
-                              color: colors.primaryColor[500],
+                              color: Colors.white,
                               fontSize: 26.sp,
                               fontWeight: FontWeight.w600),
                         ),